Moves test to new print_flags_test.go
