+++ pkg/printers/internalversion/additional_printers_test.go
@@ -26,14 +26,12 @@ import (
 	yaml "gopkg.in/yaml.v2"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	metav1beta1 "k8s.io/apimachinery/pkg/apis/meta/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	yamlserializer "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
 	"k8s.io/apimachinery/pkg/util/diff"
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
 	genericprinters "k8s.io/cli-runtime/pkg/printers"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 	"k8s.io/kubernetes/pkg/api/testapi"
@@ -199,62 +197,6 @@ func TestPrintHandlerError(t *testing.T) {
 	_, err := generator.GenerateTable(&obj, printers.GenerateOptions{})
 	if err == nil || err.Error() != "ErrorPrintHandler error" {
 		t.Errorf("Did not get the expected error: %#v", err)
-	}
-}
-
-func TestNamePrinter(t *testing.T) {
-	tests := map[string]struct {
-		obj    runtime.Object
-		expect string
-	}{
-		"singleObject": {
-			&v1.Pod{
-				TypeMeta: metav1.TypeMeta{
-					Kind: "Pod",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "foo",
-				},
-			},
-			"pod/foo\n"},
-		"List": {
-			&unstructured.UnstructuredList{
-				Object: map[string]interface{}{
-					"kind":       "List",
-					"apiVersion": "v1",
-				},
-				Items: []unstructured.Unstructured{
-					{
-						Object: map[string]interface{}{
-							"kind":       "Pod",
-							"apiVersion": "v1",
-							"metadata": map[string]interface{}{
-								"name": "bar",
-							},
-						},
-					},
-				},
-			},
-			"pod/bar\n"},
-	}
-
-	printFlags := genericclioptions.NewPrintFlags("").WithTypeSetter(legacyscheme.Scheme).WithDefaultOutput("name")
-	printer, err := printFlags.ToPrinter()
-	if err != nil {
-		t.Fatalf("unexpected err: %v", err)
-	}
-
-	for name, item := range tests {
-		buff := &bytes.Buffer{}
-		err := printer.PrintObj(item.obj, buff)
-		if err != nil {
-			t.Errorf("%v: unexpected err: %v", name, err)
-			continue
-		}
-		got := buff.String()
-		if item.expect != got {
-			t.Errorf("%v: expected %v, got %v", name, item.expect, got)
-		}
 	}
 }
