+++ staging/src/k8s.io/cli-runtime/pkg/genericclioptions/BUILD
@@ -60,11 +60,15 @@ go_test(
         "json_yaml_flags_test.go",
         "jsonpath_flags_test.go",
         "name_flags_test.go",
+        "print_flags_test.go",
         "template_flags_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
+        "//staging/src/k8s.io/client-go/kubernetes/scheme:go_default_library",
     ],
 )
